+++ src/main/java/nextstep/subway/domain/StationRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
-    @Override
     List<Station> findAll();
 }
